Import thorugh r7 (6a9d95173bcc) from upstream.

+++ BBGE/Base.cpp
@@ -22,7 +22,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 #include "Core.h"
 
 #ifdef BBGE_BUILD_WINDOWS
-	#include "ShellApi.h"
+	#include <shellapi.h>
 #endif
 
 #if defined(BBGE_BUILD_UNIX)
@@ -135,9 +135,9 @@ std::string removeSpaces(const std::string &input)
 	return result;
 }
 
-unsigned long hash(const std::string &string)
+unsigned hash(const std::string &string)
 {
-    unsigned long hash = 5381;
+    unsigned hash = 5381;
 
     for (int i = 0; i < string.size(); i++)
         hash = ((hash << 5) + hash) + string[i];
@@ -499,7 +499,7 @@ std::string stripEndlineForUnix(const std::string &in)
 	return out;
 }
 
-void forEachFile(std::string path, std::string type, void callback(const std::string &filename, int param), int param)
+void forEachFile(std::string path, std::string type, void callback(const std::string &filename, intptr_t param), intptr_t param)
 {
 	if (path.empty()) return;
